+++ test_runner/src/main/kotlin/ftl/args/IArgs.kt
@@ -21,6 +21,7 @@ interface IArgs {
     val shardTime: Int
     val repeatTests: Int
     val smartFlankGcsPath: String
+    val smartFlankDisableUpload: Boolean
     val testTargetsAlwaysRun: List<String>
     val filesToDownload: List<String>
     val disableSharding: Boolean