+++ test_runner/src/main/kotlin/ftl/gc/GcStorage.kt
@@ -54,7 +54,7 @@ object GcStorage {
         )
 
     fun uploadJunitXml(testResult: JUnitTestResult, args: IArgs) {
-        if (args.smartFlankGcsPath.isEmpty()) return
+        if (args.smartFlankGcsPath.isEmpty() || args.smartFlankDisableUpload) return
 
         // bucket/path/to/object
         val rawPath = args.smartFlankGcsPath.drop(GCS_PREFIX.length)